pyproject.toml
[tool.poetry]
name = "drf-common-exceptions"
version = "0.1.0"
description = "Common exception for Django REST framework"
authors = ["Andrey Bogoyavlensky <abogoyavlensky@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
djangorestframework = "^3.9"
django = "^2.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
isort = "^4.3"
autoflake = "^1.2"
black = "=18.9b0"
pytest-mock = "^1.10"
coverage = "^4.5"
pytest-cov = "^2.6"
pytest-watch = "^4.2"
mypy = "^0.670.0"
xenon = "^0.5.5"
pytest-django = "^3.4"

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
